refactor(#3): move login functions to bsky module
